style: rename contact me to socials
